requirements.txt
langchain>=0.1.17
langchain_ollama>=0.1.0
langchain_community>=0.0.27
langchain-huggingface>=0.0.6
langchain-chroma>=0.0.5
pymupdf
pytesseract
Pillow
chromadb
sentence-transformers

markdown2
beautifulsoup4
python-docx

docx2pdf

pandas
openpyxl
